Combine `SkinSprite` into `SkinnableSprite`
