+++ osu.Game/Skinning/DefaultSkin.cs
@@ -8,6 +8,7 @@ using osu.Framework.Audio.Sample;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.OpenGL.Textures;
+using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.Textures;
 using osu.Game.Audio;
 using osu.Game.Beatmaps.Formats;
@@ -156,6 +157,9 @@ namespace osu.Game.Skinning
 
                     break;
             }
+
+            if (GetTexture(component.LookupName) is Texture t)
+                return new Sprite { Texture = t };
 
             return null;
         }
